==> 1. HENet Model.ipynb <==
import os
from onekey_algo.classification25d.run_classification import main as clf_main
from collections import namedtuple

# 设置参数
save_dir = r'D:\x'
train_f = 'features/train_l3.txt'
val_f = 'features/val_l3.txt'
labels_f = 'features/labels.txt'
data_pattern = ''

for model_name in ['resnet50', 'resnet101', 'resnet18']:
    params = dict(train=train_f,
                  valid=val_f,
                  labels_file=labels_f,
                  data_pattern=data_pattern,
                  j=0,            
                  fname_len=3,
                  slice_size=3, 
                  clinic_size=17,
                  e2e_comp=1024,
                  max2use=None,
                  val_max2use=None,
                  batch_balance=False,
                  normalize_method='imagenet',
                  model_name=model_name,
                  gpus=[0],
                  batch_size=16,
                  epochs=90,
                  init_lr=0.01,
                  optimizer='sgd',
                  retrain=None,
                  model_root=os.path.join(save_dir, 'label3/models2.5D'),
                  iters_start=0,
                  add_date=False,
                  iters_verbose=10,
                  save_per_epoch=False,
                  pretrained=True)
    # 训练模型
    Args = namedtuple("Args", params)
    clf_main(Args(**params))
